wrote addProduct function
