--- admin/admin.js
@@ -1,5 +1,0 @@
-import { getFromLocalStorage, getLocalStorageDrinks } from "../utils";
-
-export function addProduct() {
-    return [];
-}
